added basic routing and cleaned up code on App.js
