+++ src/components/event-container/event-container.component.jsx
@@ -1,5 +1,4 @@
 import "./event-container.styles.css";
-import Button from "@mui/material/Button";
 import Stack from "@mui/material/Stack";
 import DeleteIcon from "@mui/icons-material/Delete";
 import IconButton from "@mui/material/IconButton";
